20200614HomeCommit

+++ SunRise_HOSP_MONITOR_Admin/SunRise_HOSP_MONITOR.Entity/SunRise_HOSP_MONITOR.Entity/HospMonitorManage/HistoricalPerRecordsEntity.cs
@@ -22,7 +22,7 @@ namespace SunRise_HOSP_MONITOR.Entity.HospMonitorManage
         /// 0:病人 1:陪护 2:访客
         /// </summary>
         /// <returns></returns>
-        public byte? nType { get; set; }
+        public int? nType { get; set; }
         /// <summary>
         /// 
         /// </summary>
